requirements.txt
discord.py>=2.5.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
rich>=13.7.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
